[Chore] 프로젝트 디렉터리 구조 수정
프로젝트 디렉터리 구조 수정
