+++ .gitmessage.txt
@@ -1,19 +1,19 @@
-# --- 제목 Type 설명---
-# 최대 50자 이하
-# 제목 첫 글자를 대문자로
-# 제목은 명령문으로
-# 제목 끝에 마침표(.) 금지
-# 제목과 본문을 한 줄 띄워 분리하기
-#
-# Feat: | 새로운 기능 추가 |
-# Fix: | 코드 수정 및 오류해결 |
-# Docs | 문서 수정, Readme 등 |
-# Test: | 테스트 코드 작성 및 추가 |
-# Refactor: | 코드 리펙토링 |
-# Init: | 프로젝트 초기 생성 |
-# Chore: | 기능 추가, 버그 수정과 같은 주요 변경사항이 아닌 잡다한 작업을 처리|
-# -------------------
-#
-# body: 작업내역에 대해서 상세하게 기술. 20 줄 이내로 작성.
-#
-# footer: # 이슈번호
+@ --- 제목 Type 설명---
+@ 최대 50자 이하
+@ 제목 첫 글자를 대문자로
+@ 제목은 명령문으로
+@ 제목 끝에 마침표(.) 금지
+@ 제목과 본문을 한 줄 띄워 분리하기
+@
+@ Feat: | 새로운 기능 추가 |
+@ Fix: | 코드 수정 및 오류해결 |
+@ Docs | 문서 수정, Readme 등 |
+@ Test: | 테스트 코드 작성 및 추가 |
+@ Refactor: | 코드 리펙토링 |
+@ Init: | 프로젝트 초기 생성 |
+@ Chore: | 기능 추가, 버그 수정과 같은 주요 변경사항이 아닌 잡다한 작업을 처리|
+@ -------------------
+@
+@ body: 작업내역에 대해서 상세하게 기술. 20 줄 이내로 작성.
+@
+@ footer: @ 이슈번호
