+++ src/main/java/kr/hhplus/be/server/ServerApplication.java
@@ -2,8 +2,10 @@ package kr.hhplus.be.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 
 @SpringBootApplication
+@EntityScan(basePackages = "kr.hhplus.be.server.domain")
 public class ServerApplication {
 
 	public static void main(String[] args) {
