+++ src/main/java/kr/hhplus/be/server/domain/token/TokenStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.controller.token;
+package kr.hhplus.be.server.domain.token;
 
 public enum TokenStatus {
