+++ src/main/java/kr/hhplus/be/server/global/config/swagger/BalanceApi.java
@@ -1,9 +1,10 @@
-package kr.hhplus.be.server.config.swagger;
+package kr.hhplus.be.server.global.config.swagger;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -12,10 +13,11 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import kr.hhplus.be.server.controller.user.PointResponse;
+import kr.hhplus.be.server.interfaces.balance.BalanceChargeRequest;
+import kr.hhplus.be.server.interfaces.balance.BalanceResponse;
 
-@Tag(name = "User API", description = "유저 관련 API")
-public interface UserApi {
+@Tag(name = "Balance API", description = "잔액 관련 API")
+public interface BalanceApi {
 
 	@Operation(summary = "[MOCK] 유저의 포인트 조회 API", description = "유저의 포인트를 조회합니다.")
 	@ApiResponses(value = {
@@ -24,7 +26,7 @@ public interface UserApi {
 			description = "유저 포인트 조회 성공",
 			content = @Content(
 				mediaType = "application/json",
-				schema = @Schema(implementation = PointResponse.class),
+				schema = @Schema(implementation = BalanceResponse.class),
 				examples = @ExampleObject(
 					name = "successResponse",
 					summary = "성공 응답 예시",
@@ -32,7 +34,7 @@ public interface UserApi {
 						{
 						  "message": "유저 포인트 조회 성공",
 						  "info": {
-						    "userId": 1,
+						    "balanceId": 1,
 						    "point": 50000
 						  }
 						}
@@ -59,7 +61,7 @@ public interface UserApi {
 		)
 	})
 	@GetMapping("/{userId}/points")
-	ResponseEntity<PointResponse> getPoint(@PathVariable("userId") long userId);
+	ResponseEntity<BalanceResponse> getPoint(@PathVariable("userId") long userId);
 
 	@Operation(summary = "[MOCK] 유저의 포인트 충전 API", description = "유저의 포인트 충전을 진행")
 	@ApiResponses(value = {
@@ -68,7 +70,7 @@ public interface UserApi {
 			description = "유저의 포인트 충전을 진행",
 			content = @Content(
 				mediaType = "application/json",
-				schema = @Schema(implementation = PointResponse.class),
+				schema = @Schema(implementation = BalanceResponse.class),
 				examples = @ExampleObject(
 					name = "successResponse",
 					summary = "성공 응답 예시",
@@ -76,7 +78,7 @@ public interface UserApi {
 						{
 						  "message": "유저 포인트 충전 성공",
 						  "info": {
-						    "userId": 1,
+						    "balanceId": 1,
 						    "point": 51000
 						  }
 						}
@@ -102,8 +104,8 @@ public interface UserApi {
 			)
 		)
 	})
-	@GetMapping("/{userId}/transactions")
-	ResponseEntity<PointResponse> charge(@PathVariable("userId") long userId,
-		@RequestParam(name = "chargePoint") long chargePoint);
+	@PutMapping("/{userId}/transactions")
+	ResponseEntity<BalanceResponse> charge(@PathVariable("userId") long userId,
+		@RequestBody BalanceChargeRequest request);
 
 }
