+++ src/main/java/kr/hhplus/be/server/interfaces/token/TokenResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.controller.token;
+package kr.hhplus.be.server.interfaces.token;
 
 public record TokenResponse(String message, Object tokenInfo) {
 }
